Add parse error to manifest fetch
