+++ collect_app/src/main/java/org/odk/collect/android/openrosa/OpenRosaResponseParser.java
@@ -1,0 +1,14 @@
+package org.odk.collect.android.openrosa;
+
+import org.jetbrains.annotations.Nullable;
+import org.kxml2.kdom.Document;
+import org.odk.collect.android.forms.FormListItem;
+import org.odk.collect.android.forms.MediaFile;
+
+import java.util.List;
+
+public interface OpenRosaResponseParser {
+
+    @Nullable List<FormListItem> parseFormList(Document document);
+    @Nullable List<MediaFile> parseManifest(Document document);
+}
